src/third_party/blink/web_tests/http/tests/cookie-store/delete-on-localhost.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Ensure that the Cookie Store API doesn't crash when used from a non-https
secure context. -->
<title>Cookie Store API: Delete a cookie on http://localhost</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

promise_test(async t => {
  await promise_rejects_js(t, TypeError,
      cookieStore.delete({
        name: 'cookie-name', value: 'cookie-value'
      }));
}, 'cookieStore.delete(name) throws an exception');

promise_test(async t => {
  await promise_rejects_js(t, TypeError,
      cookieStore.delete('cookie-name', 'cookie-value'));
}, `cookieStore.delete(name, value, options) throws an exception`);

</script>
